Restructure Postbox files
