+++ lib/postbox/component/component.php
