+++ lib/postbox/postbox.php
@@ -1,78 +1,13 @@
 <?php
 /**
- * File for Seravo Postbox and Toolpage.
+ * File for Seravo Postbox.
  */
 
 namespace Seravo\Postbox;
 
-use \Seravo\Helpers;
-
 // Deny direct access to this file
 if ( ! defined('ABSPATH') ) {
   die('Access denied!');
-}
-
-if ( ! class_exists('Toolpage') ) {
-  class Toolpage {
-
-    /**
-     * @var string Admin screen id where the page should be displayed in.
-     */
-    private $screen;
-
-    /**
-     * @var \Seravo\Postbox\Postbox[] Postboxes registered on the page.
-     */
-    private $postboxes = array();
-
-    /**
-     * Constructor for Toolpage. Will be called on new instance.
-     * @param string $screen Admin screen id where the page should be displayed in.
-     */
-    public function __construct( $screen = '' ) {
-      $this->screen = $screen;
-    }
-
-    /**
-     * Enables AJAX features for this page. This must be called
-     * on the page if there's even a single postbox using AJAX.
-     */
-    public function enable_ajax() {
-      // Generates WordPress nonce for this page
-      // and prints it as JavaScipt variable inside <SCRIPT>.
-      add_action(
-        'before_seravo_postboxes_' . $this->screen,
-        function() {
-          $nonce = wp_create_nonce($this->screen);
-          echo "<script>SERAVO_AJAX_NONCE = \"{$nonce}\";</script>";
-        }
-      );
-    }
-
-    /**
-     * Register postbox to be shown on the page. The same postbox
-     * instance shouldn't be used elsewhere without clone.
-     * @param \Seravo\Postbox\Postbox $postbox Postbox to be registered.
-     */
-    public function register_postbox( Postbox $postbox ) {
-      $postbox->on_page_assign($this->screen);
-      $this->postboxes[] = $postbox;
-    }
-
-    /**
-     * Register the page to be rendered. This should be called once
-     * the page is ready and all the postboxes are added.
-     * @param \Seravo\Postbox\Postbox $postbox
-     */
-    public function register_page() {
-      foreach ( $this->postboxes as $postbox ) {
-        if ( $postbox->_is_allowed() ) {
-          seravo_add_postbox($this->screen, $postbox);
-        }
-      }
-    }
-
-  }
 }
 
 if ( ! class_exists('Postbox') ) {
@@ -342,241 +277,4 @@ if ( ! class_exists('Postbox') ) {
   }
 }
 
-if ( ! class_exists('Requirements') ) {
-  final class Requirements {
-
-    /**
-     * @var string Key for 'init_from_array' initilization of is_admin.
-     */
-    const IS_ADMIN = 'is_admin';
-    /**
-     * @var string Key for 'init_from_array' initilization of is_wp_cli.
-     */
-    const IS_WP_CLI = 'is_wp_cli';
-    /**
-     * @var string Key for 'init_from_array' initilization of is_multisite.
-     */
-    const IS_MULTISITE = 'is_multisite';
-    /**
-     * @var string Key for 'init_from_array' initilization of is_not_multisite.
-     */
-    const IS_NOT_MULTISITE = 'is_not_multisite';
-    /**
-     * @var string Key for 'init_from_array' initilization of can_be_production.
-     */
-    const CAN_BE_PRODUCTION = 'can_be_production';
-    /**
-     * @var string Key for 'init_from_array' initilization of can_be_staging.
-     */
-    const CAN_BE_STAGING = 'can_be_staging';
-    /**
-     * @var string Key for 'init_from_array' initilization of can_be_development.
-     */
-    const CAN_BE_DEVELOPMENT = 'can_be_development';
-    /**
-     * @var string Key for 'init_from_array' initilization of can_be_*.
-     */
-    const CAN_BE_ANY_ENV = 'can_be_any_env';
-    /**
-     * @var string Key for 'init_from_array' initilization of capabilities.
-     */
-    const CAPABILITIES = 'capabilities';
-
-
-    /**
-     * @var bool Whether user must be admin and able to manager network.
-     */
-    public $is_admin = \true;
-    /**
-     * @var bool Whether plugin must be loaded by WP CLI.
-     */
-    public $is_wp_cli = \false;
-    /**
-     * @var bool Whether site must be multisite install.
-     */
-    public $is_multisite = \false;
-    /**
-     * @var bool Whether site must be multisite install or not.
-     */
-    public $is_not_multisite = \false;
-    /**
-     * @var bool Whether the site can be in production environment.
-     */
-    public $can_be_production = \false;
-    /**
-     * @var bool Whether the site can be in staging environment.
-     */
-    public $can_be_staging = \false;
-    /**
-     * @var bool Whether the site can be in local development environment.
-     */
-    public $can_be_development = \false;
-
-    /**
-     * Capabilities array contains capabilities as strings and
-     * arrays with capability string at index[0] and extra args at index[1].
-     * @var mixed[] Additional WordPress capabilities required.
-     * @see https://wordpress.org/support/article/roles-and-capabilities
-     */
-    public $capabilities = array();
-
-    /**
-     * Initialize requirements from array. Array should
-     * be in "[Requirements::*] => value" format.
-     * @param array<string, mixed> $requirements Requirements to be initialized.
-     */
-    public function init_from_array( $requirements ) {
-      if ( isset($requirements[self::IS_ADMIN]) ) {
-        $this->is_admin = $requirements[self::IS_ADMIN];
-      }
-      if ( isset($requirements[self::IS_WP_CLI]) ) {
-        $this->is_wp_cli = $requirements[self::IS_WP_CLI];
-      }
-      if ( isset($requirements[self::IS_MULTISITE]) ) {
-        $this->is_multisite = $requirements[self::IS_MULTISITE];
-      }
-      if ( isset($requirements[self::IS_NOT_MULTISITE]) ) {
-        $this->is_not_multisite = $requirements[self::IS_NOT_MULTISITE];
-      }
-      if ( isset($requirements[self::CAN_BE_PRODUCTION]) ) {
-        $this->can_be_production = $requirements[self::CAN_BE_PRODUCTION];
-      }
-      if ( isset($requirements[self::CAN_BE_STAGING]) ) {
-        $this->can_be_staging = $requirements[self::CAN_BE_STAGING];
-      }
-      if ( isset($requirements[self::CAN_BE_DEVELOPMENT]) ) {
-        $this->can_be_development = $requirements[self::CAN_BE_DEVELOPMENT];
-      }
-      if ( isset($requirements[self::CAN_BE_ANY_ENV]) ) {
-        $this->can_be_production = $requirements[self::CAN_BE_ANY_ENV];
-        $this->can_be_staging = $requirements[self::CAN_BE_ANY_ENV];
-        $this->can_be_development = $requirements[self::CAN_BE_ANY_ENV];
-      }
-      if ( isset($requirements[self::CAPABILITIES]) ) {
-        $this->capabilities = $requirements[self::CAPABILITIES];
-      }
-    }
-
-    /**
-     * Check if the environment and user
-     * permissions match the required ones.
-     * @return bool Whether requirements match or not.
-     */
-    public function is_allowed() {
-      if ( $this->is_admin && ! current_user_can('administrator') ) {
-        return false;
-      }
-      if ( $this->is_admin && is_multisite() && ! current_user_can('manage_network') ) {
-        return false;
-      }
-      if ( $this->is_wp_cli && ! (defined('WP_CLI') && WP_CLI) ) {
-        return false;
-      }
-      if ( $this->is_multisite && ! is_multisite() ) {
-        return false;
-      }
-      if ( $this->is_not_multisite && is_multisite() ) {
-        return false;
-      }
-      if ( ! $this->can_be_production && Helpers::is_production() ) {
-        return false;
-      }
-      if ( ! $this->can_be_staging && Helpers::is_staging() ) {
-        return false;
-      }
-      if ( ! $this->can_be_development && Helpers::is_development() ) {
-        return false;
-      }
-
-      foreach ( $this->capabilities as $capability ) {
-        $args = null;
-
-        if ( is_array($capability) ) {
-          $args = isset($capability[1]) ? $capability[1] : null;
-          $capability = $capability[0];
-        }
-
-        if ( ! \current_user_can($capability, $args) ) {
-          return false;
-        }
-      }
-
-      return \true;
-    }
-
-  }
-}
-
-namespace Seravo\Postbox\Postboxes;
-
-use \Seravo\Postbox\Ajax;
-use \Seravo\Postbox\Component;
-use \Seravo\Postbox\Template;
-
-/**
- * Pre-made Postbox for automatically executing a single
- * command and showing the output. Uses Ajax_Auto_Command
- * as the only AJAX handler.
- *
- * Component structure
- * - info paragraph
- * - command output
- */
-if ( ! class_exists('Postbox_Auto_Command') ) {
-  class Postbox_Auto_Command extends \Seravo\Postbox\Postbox {
-
-    /**
-     * @var string $info Info text to display.
-     */
-    private $info = '';
-
-    /**
-     * Constructor for Postbox_Auto_Command. Will be called on new instance.
-     * @param string $id      Unique id/slug of the postbox.
-     * @param string $context Default admin dashboard context where the postbox should be displayed in.
-     */
-    public function __construct( $id, $context = 'normal' ) {
-      parent::__construct($id, $context);
-
-      $this->set_build_func(
-        function ( Component $base ) {
-          return $this->build_auto_command($base);
-        }
-      );
-
-      $ajax_handler = new Ajax\Ajax_Auto_Command($id);
-      $this->add_ajax_handler($ajax_handler);
-    }
-
-    /**
-     * Postbox will be built here.
-     * @param \Seravo\Postbox\Component @base Base component to build postbox on.
-     */
-    public function build_auto_command( Component $base ) {
-      $base->add_child(Template::paragraph($this->info));
-      $base->add_child($this->ajax_handlers[$this->id]->get_component());
-    }
-
-    /**
-     * Set info text to display.
-     * @param string $text Info text.
-     */
-    public function set_info_text( $text ) {
-      $this->info = $text;
-    }
-
-    /**
-     * Configure the command to be executed.
-     * @param string $command Command to be executed.
-     * @param int $cache_time Seconds to cache response for (default is 300).
-     * @param bool $allow_failure Whether exit code other than 0 should respond with an error.
-     */
-    public function set_auto_command( $command, $cache_time = 300, $allow_failure = false ) {
-      $ajax_handler = $this->ajax_handlers[$this->id];
-      $ajax_handler->set_command($command);
-      $ajax_handler->allow_failure($allow_failure);
-      $ajax_handler->set_cache_time($cache_time);
-    }
-
-  }
-}
+require_once dirname(__FILE__) . '/templates/auto-command.php';
